+++ IPHashBalancer/ILoadBalancer.cs
@@ -1,0 +1,7 @@
+namespace IPHashBalancer
+{
+    public interface ILoadBalancer<T>
+    {
+        T GetNext(string key);
+    }
+}